+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.21.18
+
+### Fixed
+
+- Align different states of "Profile verifizieren" tile
+
+## 1.21.17
+
+### Fixed
+
+- Improve trust onboarding wizard route validation to prevent bypassing required steps
+
 ## 1.21.16
 
 ### Fixed